Move policy lifecycle aggregation out of bootstrap
